requirements.txt
streamlit==1.34.0
transformers==4.40.0
torch==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.4.2
joblib==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
pyvi==0.1.1
unidecode==1.3.8
